refactor: migrate from JavaScript to TypeScript
