--- src/App.jsx
@@ -1,14 +1,0 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import HomePage from './pages/HomePage'
-import LoginPage from './pages/LoginPage'
-import MyBookingsPage from './pages/MyBookingsPage'
-
-function App() {
-  return (
-    <>
-      Welcome to FlyMe
-    </>
-  )
-}
-
-export default App
