+++ src/App.tsx
@@ -1,0 +1,18 @@
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import HomePage from "./pages/HomePage";
+import LoginPage from "./pages/LoginPage";
+import MyBookingsPage from "./pages/MyBookingsPage";
+
+function App() {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/my-bookings" element={<MyBookingsPage />} />
+      </Routes>
+    </Router>
+  );
+}
+
+export default App;
