--- src/pages/HomePage.jsx
@@ -1,5 +1,0 @@
-function HomePage() {
-  return <div>HomePage works!</div>
-}
-
-export default HomePage
