+++ src/pages/HomePage.tsx
@@ -1,0 +1,5 @@
+function HomePage() {
+  return <div>HomePage works!</div>;
+}
+
+export default HomePage;
