--- src/pages/LoginPage.jsx
@@ -1,5 +1,0 @@
-function LoginPage() {
-  return <div>LoginPage works!</div>
-}
-
-export default LoginPage
