+++ src/pages/LoginPage.tsx
@@ -1,0 +1,5 @@
+function LoginPage() {
+  return <div>LoginPage works!</div>;
+}
+
+export default LoginPage;
