--- src/pages/MyBookingsPage.jsx
@@ -1,5 +1,0 @@
-function MyBookingsPage() {
-  return <div>MyBookingsPage works!</div>
-}
-
-export default MyBookingsPage
