+++ src/pages/MyBookingsPage.tsx
@@ -1,0 +1,5 @@
+function MyBookingsPage() {
+  return <div>MyBookingsPage works!</div>;
+}
+
+export default MyBookingsPage;
